fix: Updated documentation and fixed report amount from int to double

+++ backend/src/main/java/com/univr/diabetes_logger/service/ReportService.java
@@ -173,7 +173,7 @@ public class ReportService implements CrudService<Report> {
       existingReport.setMedicine(medicine);
     }
 
-    Integer amount = report.getAmount();
+    Double amount = report.getAmount();
     if (amount != null) {
       existingReport.setAmount(amount);
     }
